% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RStudio_addins.R
\name{correct_block2rchunk_Addin}
\alias{correct_block2rchunk_Addin}
\title{[!] Correnct md text genereted with Writage (add-in): block quote to R chunk}
\usage{
correct_block2rchunk_Addin()
}
\description{
[!] Correnct md text genereted with Writage (add-in): block quote to R chunk
}
\seealso{
Other 'Replace selected symbols' addins: \code{\link{correct_writage_md_addin}},
  \code{\link{rekasius2rmd_Addin}}
}
