% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RStudio_addins.R
\name{enclose_with_r_chunk_Addin}
\alias{enclose_with_r_chunk_Addin}
\title{Enclose with R code chunk}
\usage{
enclose_with_r_chunk_Addin()
}
\description{
Call this function as an addin to enclose text into R code chunk
}
