% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fread_fragmented.R
\name{fread_fragmented}
\alias{fread_fragmented}
\title{Read fragmented table from a flat file (DOES NOT WORK).}
\usage{
fread_fragmented(x, stringsAsFactors = TRUE, ...)
}
\arguments{
\item{x}{(character) Either a path to a file, a connection,
or literal data (either a single string or a raw vector).\cr
Literal data is most useful for examples and tests.
It must contain at least one new line to be recognised as data
(instead of a path).}
}
\value{
A data frame.
}
\description{
A convenience function, that wraps \code{fread}
 from package \pkg{data.table}.
}
\examples{
\dontrun{\donttest{

# Contents of file "data-raw/CO_head.txt":

#   Plant   Type  Treatment
# 1   Qn1 Quebec nonchilled
# 2   Qn1 Quebec nonchilled
# 3   Qn1 Quebec nonchilled
#   conc uptake
# 1   95   16.0
# 2  175   30.4
# 3  250   34.8


fread_fragmented("data-raw/CO_head.txt")

##   Plant   Type  Treatment conc uptake
## 1   Qn1 Quebec nonchilled   95   16.0
## 2   Qn1 Quebec nonchilled  175   30.4
## 3   Qn1 Quebec nonchilled  250   34.8
##
## Warning messages:
##     1: Missing column names filled in: 'X1' [1]
##     2: Missing column names filled in: 'X1' [1]

# These warning messages are expected.

# ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~

rez <- read_fragmented_table(
"   Plant   Type  Treatment
1   Qn1 Quebec nonchilled
2   Qn1 Quebec nonchilled
3   Qn1 Quebec nonchilled
  conc uptake
1   95   16.0
2  175   30.4
3  250   34.8")

rez
#   Plant   Type  Treatment conc uptake
# 1   Qn1 Quebec nonchilled   95   16.0
# 2   Qn1 Quebec nonchilled  175   30.4
# 3   Qn1 Quebec nonchilled  250   34.8
}}


}
