% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RStudio_addins.R
\name{insert_chunk_python_Addin}
\alias{insert_chunk_python_Addin}
\title{Insert python's code chunk for conspects}
\usage{
insert_chunk_python_Addin()
}
\description{
Call this function as an addin to insert python's code chunk for conspects at the cursor position.
}
\seealso{
Other 'Insert at cursor position' addins: 
\code{\link{insert_chunk_r_Addin}()},
\code{\link{insert_lt_quotes_low_Addin}()},
\code{\link{insert_lt_quotes_upp_Addin}()}
}
\concept{'Insert at cursor position' addins}
