% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_spAddins.R
\name{insert_lt_quotes_low_Addin}
\alias{insert_lt_quotes_low_Addin}
\title{Insert Lithuanian quotes (lower)}
\usage{
insert_lt_quotes_low_Addin()
}
\description{
Call this function as an addin to insert lower Lithuanian quotes.
}
\seealso{
Other 'Insert at cursor position' addins: \code{\link{insert_chunk_python_Addin}},
  \code{\link{insert_chunk_r_Addin}},
  \code{\link{insert_lt_quotes_upp_Addin}}
}
