% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RStudio_addins.R
\name{insert_lt_quotes_upp_Addin}
\alias{insert_lt_quotes_upp_Addin}
\title{Insert Lithuanian quotes (upper)}
\usage{
insert_lt_quotes_upp_Addin()
}
\description{
Call this function as an addin to insert upper Lithuanian quotes.
}
\seealso{
Other 'Insert at cursor position' addins: \code{\link{insert_chunk_python_Addin}},
  \code{\link{insert_chunk_r_Addin}},
  \code{\link{insert_lt_quotes_low_Addin}}
}
