% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_fragmented_table.R
\name{read_fragmented_table}
\alias{read_fragmented_table}
\title{Read franmented teble from a flat file.}
\usage{
read_fragmented_table(FILE)
}
\arguments{
\item{FILE}{(string) a file name.}
}
\value{
a data table.
}
\description{
A convenience function, that wraps \code{read_tabe}
 from package \pkg{readr}.
}
\examples{
\dontrun{\donttest{

# Contents of file "data-raw/CO_head.txt":

#   Plant   Type  Treatment
# 1   Qn1 Quebec nonchilled
# 2   Qn1 Quebec nonchilled
# 3   Qn1 Quebec nonchilled
#   conc uptake
# 1   95   16.0
# 2  175   30.4
# 3  250   34.8


read_fragmented_table("data-raw/CO_head.txt")

##   Plant   Type  Treatment conc uptake
## 1   Qn1 Quebec nonchilled   95   16.0
## 2   Qn1 Quebec nonchilled  175   30.4
## 3   Qn1 Quebec nonchilled  250   34.8
##
## Warning messages:
##     1: Missing column names filled in: 'X1' [1]
##     2: Missing column names filled in: 'X1' [1]

# These warning messages are expected.

}}

}

