% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RStudio_addins.R
\name{rekasius2rmd_Addin}
\alias{rekasius2rmd_Addin}
\title{[!] Correnct md text genereted with Writage (add-in): block quote to R chunk}
\usage{
rekasius2rmd_Addin()
}
\description{
[!] Correnct md text genereted with Writage (add-in): block quote to R chunk
}
\seealso{
Other 'Replace selected symbols' addins: \code{\link{correct_block2rchunk_Addin}},
  \code{\link{correct_writage_md_addin}}
}
