% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RStudio_addins.R
\name{correct_writage_md_addin}
\alias{correct_writage_md_addin}
\title{[!] Correnct md text genereted with Writage (add-in)}
\usage{
correct_writage_md_addin()
}
\description{
[!] Correnct md text genereted with Writage (add-in)
}
\seealso{
Other 'Replace selected symbols' addins: \code{\link{correct_block2rchunk_Addin}},
  \code{\link{rekasius2rmd_Addin}}
}
